Add ChartContainer to SimpleNormalizedChart
